# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bluntly.ph API",
    "description": "Verified product & seller review platform. All error responses use the RFC 9457 problem+json schema. Identity is a FastAPI-issued JWT (ADR-010).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/review-queue": {
      "get": {
        "tags": [
          "admin: referral"
        ],
        "summary": "Moderator queue: pending reviews + monetized-but-edited",
        "operationId": "review_queue_api_v1_admin_review_queue_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewQueueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/reviews/{review_id}/referral-link": {
      "post": {
        "tags": [
          "admin: referral"
        ],
        "summary": "Paste referral link -> monetize + publish (atomic)",
        "operationId": "attach_link_api_v1_admin_reviews__review_id__referral_link_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachLinkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin: referral"
        ],
        "summary": "Revoke the active referral link (stays published)",
        "operationId": "revoke_link_api_v1_admin_reviews__review_id__referral_link_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/reviews/{review_id}/publish": {
      "post": {
        "tags": [
          "admin: referral"
        ],
        "summary": "Publish without a link (<=2* -> Honesty Fund, else approved)",
        "operationId": "publish_api_v1_admin_reviews__review_id__publish_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/reviews/{review_id}/reject": {
      "post": {
        "tags": [
          "admin: referral"
        ],
        "summary": "Reject a queued review (stays hidden; author may resubmit)",
        "operationId": "reject_api_v1_admin_reviews__review_id__reject_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/reviews/{review_id}/unpublish": {
      "post": {
        "tags": [
          "admin: referral"
        ],
        "summary": "Take a published review off the public site",
        "operationId": "unpublish_api_v1_admin_reviews__review_id__unpublish_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptionalReasonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/reviews/{review_id}/referral-links": {
      "get": {
        "tags": [
          "admin: referral"
        ],
        "summary": "Referral link history for a review",
        "operationId": "link_history_api_v1_admin_reviews__review_id__referral_links_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferralLinkOut"
                  },
                  "title": "Response Link History Api V1 Admin Reviews  Review Id  Referral Links Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/commissions/import": {
      "post": {
        "tags": [
          "admin: earnings"
        ],
        "summary": "Import a monthly commission CSV (all-or-nothing, idempotent)",
        "operationId": "import_commissions_api_v1_admin_commissions_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_commissions_api_v1_admin_commissions_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/honesty-fund/run": {
      "post": {
        "tags": [
          "admin: earnings"
        ],
        "summary": "Run the Honesty Fund distribution for a cycle (idempotent)",
        "operationId": "run_honesty_fund_api_v1_admin_honesty_fund_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HonestyFundRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HonestyFundRunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttachLinkRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          }
        },
        "type": "object",
        "required": [
          "url",
          "platform"
        ],
        "title": "AttachLinkRequest"
      },
...
      "Body_import_commissions_api_v1_admin_commissions_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_commissions_api_v1_admin_commissions_import_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HonestyFundRunRequest": {
        "properties": {
          "cycle_month": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}-\\d{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cycle Month"
          }
        },
        "type": "object",
        "title": "HonestyFundRunRequest"
      },
      "HonestyFundRunResult": {
        "properties": {
          "cycle_month": {
            "type": "string",
            "format": "date",
            "title": "Cycle Month"
          },
          "pool": {
            "type": "string",
            "title": "Pool"
          },
          "recipients": {
            "type": "integer",
            "title": "Recipients"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "cycle_month",
          "pool",
          "recipients",
          "status"
        ],
        "title": "HonestyFundRunResult"
      },
      "ImportResult": {
        "properties": {
          "imported": {
            "type": "integer",
            "title": "Imported"
          },
          "skipped_duplicates": {
            "type": "integer",
            "title": "Skipped Duplicates"
          },
          "unmatched": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Unmatched"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          }
        },
        "type": "object",
        "required": [
          "imported",
          "skipped_duplicates",
          "unmatched",
          "total_rows"
        ],
        "title": "ImportResult"
      },
...
      "OptionalReasonRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "OptionalReasonRequest"
      },
      "Platform": {
        "type": "string",
        "enum": [
          "shopee",
          "lazada",
          "amazon",
          "other"
        ],
        "title": "Platform"
      },
...
      "QueueItem": {
        "properties": {
          "review": {
            "$ref": "#/components/schemas/ReviewOut"
          },
          "product": {
            "$ref": "#/components/schemas/QueueProduct"
          },
          "author": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueueAuthor"
              },
              {
                "type": "null"
              }
            ]
          },
          "suggested_platform": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Platform"
              },
              {
                "type": "null"
              }
            ]
          },
          "edited_since_monetized": {
            "type": "boolean",
            "title": "Edited Since Monetized",
            "default": false
          },
          "signals": {
            "$ref": "#/components/schemas/QueueSignals"
          }
        },
        "type": "object",
        "required": [
          "review",
          "product"
        ],
        "title": "QueueItem"
      },
...
      "ReasonRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "ReasonRequest"
      },
      "ReferralLinkOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "review_id": {
            "type": "string",
            "format": "uuid",
            "title": "Review Id"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "status": {
            "$ref": "#/components/schemas/ReferralLinkStatus"
          },
          "review_version": {
            "type": "integer",
            "title": "Review Version"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "revoked_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked By"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "revoke_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoke Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "review_id",
          "platform",
          "url",
          "status",
          "review_version",
          "created_at"
        ],
        "title": "ReferralLinkOut"
      },
...
      "ReviewOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "review_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Id"
          },
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "author_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "discussion": {
            "type": "string",
            "title": "Discussion"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          },
          "verdict_explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict Explanation"
          },
          "target_audience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Audience"
          },
          "anti_target_audience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anti Target Audience"
          },
          "star_rating": {
            "type": "integer",
            "title": "Star Rating"
          },
          "pros": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pros"
          },
          "cons": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cons"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "receipt_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receipt Url"
          },
          "price_paid": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Paid"
          },
          "verification_status": {
            "$ref": "#/components/schemas/VerificationStatus"
          },
          "helpful_votes": {
            "type": "integer",
            "title": "Helpful Votes",
            "default": 0
          },
          "unhelpful_votes": {
            "type": "integer",
            "title": "Unhelpful Votes",
            "default": 0
          },
          "wilson_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Wilson Score",
            "default": "0"
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "earn_eligible_status": {
            "$ref": "#/components/schemas/EarnEligibleStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "referral_redirect_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Redirect Url",
            "description": "Public attribution link \u2014 exposed instead of the raw affiliate URL.\nOnly present once the review is published AND monetized.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_id",
          "title",
          "discussion",
          "verdict",
          "star_rating",
          "verification_status",
          "current_version",
          "earn_eligible_status",
          "created_at",
          "updated_at",
          "referral_redirect_url"
        ],
        "title": "ReviewOut"
      },
      "ReviewQueueResponse": {
        "properties": {
          "pending": {
            "items": {
              "$ref": "#/components/schemas/QueueItem"
            },
            "type": "array",
            "title": "Pending"
          },
          "edited_since_monetized": {
            "items": {
              "$ref": "#/components/schemas/QueueItem"
            },
            "type": "array",
            "title": "Edited Since Monetized"
          }
        },
        "type": "object",
        "required": [
          "pending",
          "edited_since_monetized"
        ],
        "title": "ReviewQueueResponse"
      },
...
      "Verdict": {
        "type": "string",
        "enum": [
          "yes_absolutely",
          "it_depends",
          "hard_pass"
        ],
        "title": "Verdict"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}